+++ src/InvestmentHub/Providers/Rico/InvestmentHub.Providers.Models.Rico/Responses/GetPositionDetailResponse.cs
@@ -4,7 +4,7 @@ namespace InvestmentHub.Providers.Models.Rico.Responses
 {
     public class GetPositionDetailResponse
     {
-        public float CurrentTotalValue { get; set; }
+        public double CurrentTotalValue { get; set; }
         public IEnumerable<PositionDetail> Positions { get; set; }
     }
 }
